+++ packages/stylelint-config-wordpress/index.js
@@ -31,6 +31,7 @@ module.exports = {
     "declaration-colon-space-after": "always-single-line",
     "declaration-colon-space-before": "never",
     "font-family-name-quotes": "always-where-recommended",
+    "font-weight-notation": "numeric",
     "function-calc-no-unspaced-operator": true,
     "function-comma-space-after": "always",
     "function-comma-space-before": "never",