+++ packages/stylelint-config-wordpress/index.js
@@ -45,6 +45,9 @@ module.exports = {
     "keyframe-declaration-no-important": true,
     "length-zero-no-unit": true,
     "max-empty-lines": 2,
+    "max-line-length": [ 80, {
+      "ignore": "non-comments",
+    } ],
     "media-feature-colon-space-after": "always",
     "media-feature-colon-space-before": "never",
     "media-feature-no-missing-punctuation": true,