+++ packages/stylelint-config-wordpress/index.js
@@ -59,7 +59,7 @@ module.exports = {
     "no-eol-whitespace": true,
     "no-extra-semicolons": true,
     "no-invalid-double-slash-comments": true,
-    "no-missing-eof-newline": true,
+    "no-missing-end-of-source-newline": true,
     "number-leading-zero": "always",
     "number-no-trailing-zeros": true,
     "property-case": "lower",