+++ packages/stylelint-config-wordpress/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Removed: `stylelint < 6.9.0` compatibility.
 * Added: `selector-attribute-quotes` rule.
 * Added: `font-weight-notation` rule.
+* Added: `max-line-length` rule.
 
 # 8.0.0