+++ packages/stylelint-config-wordpress/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added: `font-weight-notation` rule.
 * Added: `max-line-length` rule.
 * Deprecated `no-missing-eof-newline` rule. Use the new `no-missing-end-of-source-newline` rule instead.
+* Fixed `font-family-name-quotes` test warning message in `values.js`.
 
 # 8.0.0