+++ packages/stylelint-config-wordpress/CHANGELOG.md
@@ -1,4 +1,4 @@
-# HEAD
+# 9.0.0
 
 * Removed: `stylelint < 7.2.0` compatibility.
 * Removed: NodeJS 0.12.x support, `stylelint` and `stylelint-config-wordpress` now require NodeJS > 4.2.1 LTS or greater